reliability-service/requirements.txt
fastapi==0.78.0
uvicorn==0.17.6
redis==4.3.4
kafka-python==2.0.2
prometheus-client==0.16.0
opentelemetry-api==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-instrumentation-redis==0.47b0
opentelemetry-instrumentation-kafka-python==0.47b0
boto3==1.34.69
